% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{column_vector}
\alias{column_vector}
\title{Get the contents of a single column of a tbl}
\usage{
column_vector(d, i)
}
\arguments{
\item{d}{a tbl}

\item{i}{a character string containing the name of the column to be
returned OR
a single number representing the position of the column to be
returned}
}
\value{
A vector containing the values from the specified column.
         The type will be numeric, factor, or character, depending on the
         type of the original column
}
\description{
Function to return a vector with the contents of one column of a tbl_df
The type of the column will be preserved if numeric or factor; all else will
be conveted to a vector of character strings
}
\examples{
x <- dplyr::tbl_df(iris)
column_vector(x,2)
column_vector(x,"Species")
}

